refac: rearrange controller docs
