+++ apps/server/src/features/category/controllers/category.controller.ts
@@ -1,7 +1,8 @@
 import { Controller, Get, HttpException, HttpStatus } from '@nestjs/common';
-import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { ApiTags } from '@nestjs/swagger';
 import { Category } from '../domain/category';
 import { CategoryService } from '../services/category.service';
+import { GetAllCategoriesDocs } from './docs/get-all-categories.swagger.decorator';
 
 @ApiTags('Categories')
 @Controller('categories')
@@ -9,35 +10,7 @@ export class CategoryController {
   constructor(private readonly categoryService: CategoryService) {}
 
   @Get()
-  @ApiOperation({
-    summary: 'Get all categories',
-    description: 'Return the list of all categories available for projects',
-  })
-  @ApiResponse({
-    status: 200,
-    description: 'List of all categories available',
-    schema: {
-      type: 'array',
-      items: {
-        type: 'object',
-        properties: {
-          id: { type: 'string', example: '1' },
-          name: { type: 'string', example: 'Développement Web' },
-        },
-      },
-    },
-  })
-  @ApiResponse({
-    status: 500,
-    description: 'Server error when fetching categories',
-    schema: {
-      type: 'object',
-      properties: {
-        message: { type: 'string', example: 'Internal server error' },
-        statusCode: { type: 'number', example: 500 },
-      },
-    },
-  })
+  @GetAllCategoriesDocs()
   async getAllCategories(): Promise<Category[]> {
     const result = await this.categoryService.getAllCategories();
     if (!result.success) {
