+++ apps/server/src/features/project/controllers/docs/update-project-by-id.swagger.decorator.ts
@@ -1,0 +1,113 @@
+import { applyDecorators } from "@nestjs/common";
+import { ApiBody, ApiOperation, ApiResponse } from "@nestjs/swagger";
+import { UpdateProjectDto } from "../dto/update-project.dto";
+
+export function UpdateProjectByIdDocs() {
+  return applyDecorators(
+  ApiOperation({
+    summary:
+      'Mettre à jour un projet, si le titre et la description changent, le repo github sera mis à jour',
+  }),
+  ApiBody({
+    type: UpdateProjectDto,
+    description: 'Données du projet à mettre à jour',
+  }),
+  ApiResponse({
+    status: 200,
+    description: 'Projet mis à jour avec succès',
+    example: {
+      id: 'e214183d-bf0a-4818-9050-17d4905de4e8',
+      ownerId: 'GCJO6PXIysuDms1Od6W8TefrigEamAeP',
+      title: 'test3',
+      description: 'nouvelle description',
+      image: '',
+      coverImages: [],
+      readme: '',
+      categories: [
+        {
+          id: '3',
+          name: 'Applications Mobile',
+        },
+        {
+          id: '5',
+          name: 'Jeux Vidéo',
+        },
+      ],
+      techStacks: [
+        {
+          id: '2',
+          name: 'Next.js',
+          iconUrl:
+            'https://cdn.jsdelivr.net/gh/devicons/devicon@latest/icons/nextjs/nextjs-original.svg',
+          type: 'TECH',
+        },
+        {
+          id: '4',
+          name: 'Vue.js',
+          iconUrl:
+            'https://cdn.jsdelivr.net/gh/devicons/devicon@latest/icons/vuejs/vuejs-original.svg',
+          type: 'TECH',
+        },
+      ],
+      projectRoles: [
+        {
+          id: 'e07b8ddb-7df8-451d-b8fb-141975e1f2ef',
+          title: 'Frontend Developer',
+          description: 'Développeur React expérimenté',
+          techStacks: [
+            {
+              id: '2',
+              name: 'Next.js',
+              iconUrl:
+                'https://cdn.jsdelivr.net/gh/devicons/devicon@latest/icons/nextjs/nextjs-original.svg',
+              type: 'TECH',
+            },
+          ],
+        },
+        {
+          id: '679f7647-c7d2-4991-b663-fe000d060327',
+          title: 'Frontend Developer',
+          description: 'Développeur React junior',
+          techStacks: [
+            {
+              id: '2',
+              name: 'Next.js',
+              iconUrl:
+                'https://cdn.jsdelivr.net/gh/devicons/devicon@latest/icons/nextjs/nextjs-original.svg',
+              type: 'TECH',
+            },
+          ],
+        },
+        {
+          id: '13353136-d190-4ee0-a79e-7f41c9122f7d',
+          title: 'Backend Developer',
+          description: 'Développeur php junior',
+          techStacks: [
+            {
+              id: '2',
+              name: 'Next.js',
+              iconUrl:
+                'https://cdn.jsdelivr.net/gh/devicons/devicon@latest/icons/nextjs/nextjs-original.svg',
+              type: 'TECH',
+            },
+          ],
+        },
+      ],
+      externalLinks: [
+        {
+          id: 'bdaf22c9-ec51-4449-b5ca-8a79f33949c2',
+          type: 'github',
+          url: 'https://github.com/lhourquin/test2',
+        },
+        {
+          id: 'b3e01acd-c7e4-4815-bb04-0764ce9a64de',
+          type: 'twitter',
+          url: 'https://x.com/kalu',
+        },
+      ],
+      createdAt: '2025-08-31T10:36:15.264Z',
+      updatedAt: '2025-08-31T17:09:13.569Z',
+    },
+  }),
+  );
+}
